refactor: separate source files
