+++ source/utils.js
@@ -1,0 +1,5 @@
+import decamelize from 'decamelize';
+
+export const decamelizeFlagKey = flagKey => `--${decamelize(flagKey, {separator: '-'})}`;
+
+export const joinFlagKeys = (flagKeys, prefix = '--') => `\`${prefix}${flagKeys.join(`\`, \`${prefix}`)}\``;
