+++ test/fixtures/fixture-required.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import meow from '../../index.js';
+import meow from '../../source/index.js';
 
 const cli = meow({
 	importMeta: import.meta,
